Change js to tsx
